Fix compiler errors/warnings in r8355
git-svn-id: http://skia.googlecode.com/svn/trunk@8357 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ debugger/SkDrawCommand.cpp
@@ -290,7 +290,7 @@ DrawData::DrawData(const void* data, size_t length) {
 
     // TODO: add display of actual data?
     SkString* str = new SkString;
-    str->appendf("length: %d", length);
+    str->appendf("length: %d", (int) length);
     fInfo.push(str);
 }
